+++ src/joedb/Writable.h
@@ -107,7 +107,7 @@ namespace joedb
 
    virtual void on_blob(Blob blob) {}
    virtual bool wants_blob_data() const {return false;}
-   virtual Blob write_blob_data(const std::string &data) {return Blob();}
+   virtual Blob write_blob(const std::string &data) {return Blob();}
 
    virtual ~Writable() = default;
  };