+++ src/joedb/ui/Blob_Reader_Command_Processor.cpp
@@ -21,7 +21,7 @@ namespace joedb
 
    if (!blob.is_null())
    {
-    const std::string s = blob_reader.read_blob_data(blob);
+    const std::string s = blob_reader.read_blob(blob);
     const std::string file_name = read_string(parameters);
 
     if (file_name.empty())