+++ src/joedb/ui/Network_Connection_Builder.h
@@ -1,4 +1,4 @@
-#include "joedb/concurrency/Network_Channel.h"
+#include "joedb/concurrency/Network_Connector.h"
 #include "joedb/concurrency/Robust_Connection.h"
 #include "joedb/concurrency/Server_File.h"
 #include "joedb/ui/Connection_Builder.h"